requirements.txt
requests==2.32.2
cachetools==5.3.2
requests==2.32.2
coloredlogs==15.0.1
setuptools==70.0.0
Jinja2==3.1.4
PyYAML==6.0.1
types-pyyaml
openstacksdk==3.3.0
pytest==7.4.0
mypy==1.4.1
flake8==6.1.0
